feat: add function by choose types of places(monument or musemium) by eventLister menubar buttons
