+++ index.html
@@ -10,14 +10,12 @@
 <body>
     <div class="conteiner">
     <div id="navbarContainer">
-        <div class="header">Карта памяти</div>
-        <div id="map">
-
-        </div>
+       
+        <div id="map"></div>
         <div id="navbar">
           <div id="bubbleWrapper">
-            <div id="bubble1" onclick= "move('1', '228px', '#ffcc80')" class="bubble"><span class="icon"><i class="fas fa-home"></i></span></div>
-            <div id="bubble2" onclick= "move('2', '682px', '#81d4fa')" class="bubble"><span class="icon"><i class="fab fa-twitter"></i></span></div>
+            <div id="bubble1" onclick= "move('1', '228px', '#ffcc80')" onclick = "changeType('all')" class="bubble"><span class="icon"><i class="fas fa-home"></i></span></div>
+            <div id="bubble2" onclick= "move('2', '682px', '#81d4fa')" onclick = "changeType('monument')" class="bubble"><span class="icon"><i class="fab fa-twitter"></i></span></div>
             <div id="bubble3" onclick= "move('3', '1138px', '#c5e1a5')"class="bubble"><span class="icon"><i class="fas fa-bell"></i></span></div>
             <div id="bubble4" onclick= "move('4', '1592px', '#ce93d8')"class="bubble"><span class="icon"><i class="fas fa-user"></i></span></div>
           </div>
@@ -39,11 +37,12 @@
             </filter>
           </defs>
         </svg>
-        <script src="https://api-maps.yandex.ru/2.1/?apikey=82cd32e7-4cca-486d-a35e-2fd405dd3d6f&lang=ru_RU"
-        type="text/javascript"></script>
+        <script src="https://api-maps.yandex.ru/2.1/?apikey=82cd32e7-4cca-486d-a35e-2fd405dd3d6f&lang=ru_RU"></script>
+        <script src = "map.js"type="module" ></script>
+       <script src = "places.js"></script>
         <script src="https://cdnjs.cloudflare.com/ajax/libs/gsap/3.9.1/gsap.min.js"></script>
         <script src = "script.js"></script>
-        <script src = "map.js" type="model"></script>
+        
 </body>
 
 </html>
