+++ places.js
@@ -1,0 +1,24 @@
+
+let person1 ={
+    name: "Журавли",
+    address: "Территория парка Победы",
+    description: "",
+    imageSrc: "placesImage/jurawli.jpg",
+    coords:[51.5442664,46.0496873],
+    metkaSrc: "sources/commonPlaceMark.png",
+    type: "monument"
+}
+
+let person2 ={
+    name: "Журавли",
+    address: "Территория парка Победы",
+    description: "",
+    imageSrc: "placesImage/jurawli.jpg",
+    coords:[51.4442664,46.0496873],
+    metkaSrc: "sources/commonPlaceMark.png",
+    type: "mesem"
+}
+
+export let listPlaces = [person1,person2]
+
+
