--- sources/places.js
@@ -1,13 +1,0 @@
-
-let person1 ={
-    name: "Журавли",
-    address: "Территория парка Победы",
-    description: "",
-    imageSrc: "placesImage/jurawli.jpg",
-    coords:[51.5442664,46.0496873],
-    metkaSrc: "sources/commonPlaceMark.png"
-}
-
-export let listPlaces = [person1];
-
-
