+++ style.css
@@ -1,5 +1,4 @@
 body {
-  background: #37474f;
   display: flex;
   justify-content: center;
   align-items: center;
@@ -7,8 +6,7 @@ body {
   overflow: hidden;
 }
 #map{
-  width: 100%;
-  height: 90%;
+  height: 100%;
 }
 #bubble1{
   width: 60px;
@@ -32,13 +30,13 @@ body {
   background-image: url("sources/home.png");
 }
 .conteiner{
-  width: 2000px;
+  width: 1900px;
 }
 
 #navbarContainer{
   
   width: 1820px;
-  height: 950px;
+  height: 920px;
   margin-left: auto;
   margin-right: auto;
   border-radius: 20px;
